refactor(core): Use schema introspection for entity ID encoding/decoding
Rather than needing to specify which fields need to be decoded via the `@Decode()` decorator, we now introspect the GraphQL schema to query the type of each field in the input & output of an operation. This solves issues such as #131 and also does away with the need to manually keep track of which fields need encoding & decoding.

Closes #131