fix(admin-ui): Add polling to watch mode to alleviate race condition
When in watch mode with UI extensions, the `ng serve` watch mode will start to re-compile as soon as the first UI extension file is copied. When there are a lot of files to copy, this leads to race conditions where the app cannot be compiled because now all the UI extension files have been copied over yet.