feat(core): Use decimal type for OrderItem.taxRate
Relates to #234.

BREAKING CHANGE: The `OrderItem.taxRate` column type in the database has been changed from `int` to `decimal`. You will need to perform a migration to update this column and depending on your database type, you may need to manually edit the migration script in order to preserve the old values.