fix(ui-devkit): Correctly handle static asset file paths
Previously, passing a the path of a file rather than a dir in the extension `staticAssets` array resulted in an uncaught error. This change fixes it so individual files as well as directories are correctly copied over to the custom admin-ui project. This makes it possible e.g. to specify your own logo images (as in #309) to overwrite the default logo images.