feat(core): Allow multiple Fulfillments per OrderItem
Relates to #565. This change means that a Fulfillment can be cancelled, and then the related
OrderItems can be assigned to a new Fulfillment.

BREAKING CHANGE: A change to the relation between OrderItems and Fulfillments means a database
migration will be required to preserve fulfillment data of existing Orders.
See the release blog post for details.